requirements.txt
fastapi==0.103.1
torch==2.0.1
transformers==4.33.3
uvicorn==0.23.2
python-multipart==0.0.6
Pillow==10.0.1